[🚚rename]: filter store 파일 위치 이동
